refactor(nn): restructure LSTM module into separate files; write tests.
